% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_null_deaths_binary}
\alias{data_null_deaths_binary}
\title{Example dataset with null treatment effects and deaths (binary outcome)}
\format{
A data table with 21,674 rows and 8 variables:
\describe{
\item{time}{Time index.}
\item{id}{Unique identifier for each individual.}
\item{L}{Binary time-varying covariate.}
\item{Z}{Medication initiated at baseline.}
\item{A}{Binary indicator of adhering to medication initiated at baseline.}
\item{R}{Indicator if the outcome of interest is measured.}
\item{Y}{Binary outcome of interest.}
\item{D}{Indicator if death occured.}
}
}
\usage{
data_null_deaths_binary
}
\description{
A dataset consisting of 21,674 observations on 1,000 individuals over 25 time points. Each row in the dataset corresponds to the record of one individual at one time point.
}
\keyword{datasets}
