% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipw.R
\name{ipw}
\alias{ipw}
\title{Pooled inverse probability weighting}
\usage{
ipw(
  data,
  pooled = TRUE,
  A_model,
  R_model_numerator,
  R_model_denominator,
  Y_model,
  truncation_percentile = NULL
)
}
\arguments{
\item{data}{Data frame containing the observed data}

\item{pooled}{Logical scalar specifying whether the pooled or nonpooled IPW method is applied. The default is \code{TRUE}, i.e., the pooled IPW method.}

\item{A_model}{Model statement for the treatment variable}

\item{R_model_numerator}{Model statement for the indicator variable for the measurement of the outcome variable, used in the numerator of the IP weights}

\item{R_model_denominator}{Model statement for the indicator variable for the measurement of the outcome variable, used in the denominator of the IP weights}

\item{Y_model}{Model statement for the outcome variable}

\item{truncation_percentile}{Numerical scalar specifying the percentile by which to truncated the IP weights}
}
\value{
A list that includes the following components:
\item{est}{A data frame containing the counterfactual mean estimates for each medication at each time interval.}
\item{args}{A list containing the arguments supplied to \code{ipw}, except the observed data set.}
}
\description{
This function applies the pooled inverse probability weighted approach described by McGrath et al. (in preparation).
}
\details{
Additional description of the method
}
\examples{
res <- ipw(data = data_null,
           pooled = TRUE,
           A_model = A ~ L + Z,
           R_model_numerator = R ~ L0 + Z,
           R_model_denominator = R ~ L + A + Z,
           Y_model = Y ~ L0 * (t0 + Z))
res$est

}
