% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{print.ipw}
\alias{print.ipw}
\title{Print method for "ipw" objects}
\usage{
\method{print}{ipw}(x, ...)
}
\arguments{
\item{x}{Object of class "ipw".}

\item{...}{Other arguments.}
}
\value{
No value is returned.
}
\description{
Print method for objects of class "ipw".
}
\examples{
data_null_processed <- prep_data(data = data_null, grace_period_length = 2,
                                 baseline_vars = 'L')
res <- ipw(data = data_null_processed,
           pooled = TRUE,
           outcome_times = c(6, 12, 18, 24),
           A_model = A ~ L + Z,
           R_model_numerator = R ~ L_baseline + Z,
           R_model_denominator = R ~ L + A + Z,
           Y_model = Y ~ L_baseline * (t0 + Z))
print(res)

}
\seealso{
\code{\link{ipw}}
}
