% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_null}
\alias{data_null}
\title{Example dataset with null treatment effects}
\format{
A data table with 25,000 rows and 7 variables:
\describe{
\item{time}{Time index.}
\item{id}{Unique identifier for each individual.}
\item{L}{Binary time-varying covariate.}
\item{Z}{Medication initiated at baseline.}
\item{A}{Binary indicator of adhering to medication initiated at baseline.}
\item{R}{Indicator if the outcome of interest is measured.}
\item{Y}{Continuous outcome of interest.}
}
}
\usage{
data_null
}
\description{
A dataset consisting of 25,000 observations on 1,000 individuals over 25 time points. Each row in the dataset corresponds to the record of one individual at one time point.
}
\keyword{datasets}
