% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{get_CI}
\alias{get_CI}
\title{Bootstrap-based confidence intervals}
\usage{
get_CI(ipw_res, data, n_boot, conf_level = 0.95)
}
\arguments{
\item{ipw_res}{Output from the \code{ipw} function.}

\item{data}{Data frame containing the observed data}

\item{n_boot}{Numeric scalar specifying the number of bootstrap replicates to use}

\item{conf_level}{Numeric scalar specifying the confidence level for the confidence intervals. The default is \code{0.95}.}
}
\value{
A list that includes the following components:
\item{res_boot}{A list where each component corresponds to a different medication \eqn{z} level. Each component of the list is a data frame containing the estimates and confidence intervals for the counterfactual outcome mean under the treatment regime indexed by \eqn{z}.}
\item{res_boot_all}{A three dimensional array containing all the bootstrap replicates. The first dimension corresponds to the bootstrap replicate; The second dimension corresponds to the time interval; The third dimension corresponds to the medication \eqn{z} level.}
}
\description{
This function applies nonparametric bootstrap to construct confidence intervals around the counterfactual mean estimates obtained by \code{\link{ipw}}.
}
\details{
Additional description of the method
}
\examples{
\donttest{
set.seed(1234)
data_null_processed <- prep_data(data = data_null, grace_period_length = 2,
                                 baseline_vars = 'L')
res_est <- ipw(data = data_null_processed,
               pooled = TRUE,
               outcome_times = c(6, 12, 18, 24),
               A_model = A ~ L + Z,
               R_model_numerator = R ~ L_baseline + Z,
               R_model_denominator = R ~ L + A + Z,
               Y_model = Y ~ L_baseline * (t0 + Z))
res_ci <- get_CI(ipw_res = res_est, data = data_null_processed, n_boot = 10)
res_ci$res_boot
}


}
