% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipw.R
\name{ipw}
\alias{ipw}
\title{Pooled inverse probability weighting}
\usage{
ipw(
  data,
  pooled = TRUE,
  outcome_times,
  A_model,
  R_model_numerator,
  R_model_denominator,
  Y_model,
  truncation_percentile = NULL,
  return_model_fits = TRUE
)
}
\arguments{
\item{data}{Data frame containing the observed data}

\item{pooled}{Logical scalar specifying whether the pooled or nonpooled IPW method is applied. The default is \code{TRUE}, i.e., the pooled IPW method.}

\item{outcome_times}{Numeric vector specifying the follow-up time(s) of interest for the counterfactual outcome mean. The default is all time points in \code{data}.}

\item{A_model}{Model statement for the treatment variable}

\item{R_model_numerator}{Model statement for the indicator variable for the measurement of the outcome variable, used in the numerator of the IP weights}

\item{R_model_denominator}{Model statement for the indicator variable for the measurement of the outcome variable, used in the denominator of the IP weights}

\item{Y_model}{Model statement for the outcome variable}

\item{truncation_percentile}{Numerical scalar specifying the percentile by which to truncated the IP weights}

\item{return_model_fits}{Logical scalar specifying whether to include the fitted models in the output}
}
\value{
A object of class "ipw". This object is a list that includes the following components:
\item{est}{A data frame containing the counterfactual mean estimates for each medication at each time interval.}
\item{model_fits}{A list containing the fitted models for the treatment, outcome measurement, and outcome (if \code{return_model_fits} is set to \code{TRUE})}
\item{args}{A list containing the arguments supplied to \code{\link{ipw}}, except the observed data set.}
}
\description{
This function applies the pooled inverse probability weighted (IPW) approach described by McGrath et al. (in preparation).
}
\details{
Additional description of the method
}
\examples{
data_null_processed <- prep_data(data = data_null, grace_period_length = 2,
                                 baseline_vars = 'L')
res <- ipw(data = data_null_processed,
           pooled = TRUE,
           outcome_times = c(6, 12, 18, 24),
           A_model = A ~ L + Z,
           R_model_numerator = R ~ L_baseline + Z,
           R_model_denominator = R ~ L + A + Z,
           Y_model = Y ~ L_baseline * (t0 + Z))
res$est

}
