% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ipw}
\alias{ipw}
\title{Pooled inverse probability weighting}
\usage{
ipw(
  df,
  pooled = TRUE,
  A_model,
  R_model_numerator,
  R_model_denominator,
  Y_model,
  truncation_percentile = NULL
)
}
\arguments{
\item{df}{Data frame containing the observed data}

\item{pooled}{Logical scalar specifying whether the pooled or nonpooled IPW method is applied. The default is \code{TRUE}, i.e., the pooled IPW method.}

\item{A_model}{Model statement for the treatment variable}

\item{R_model_numerator}{Model statement for the indicator variable for the measurement of the outcome variable, used in the numerator of the IP weights}

\item{R_model_denominator}{Model statement for the indicator variable for the measurement of the outcome variable, used in the denominator of the IP weights}

\item{Y_model}{Model statement for the outcome variable}

\item{truncation_percentile}{Numerical scalar specifying the percentile by which to truncated the IP weights}
}
\value{
A list that includes the following components:
\item{est_z0}{A vector containing the counterfactual mean estimates under medication \eqn{Z=0}. The \eqn{k}th element in the vector corresponds to the \eqn{k}th follow-up time interval.}
\item{est_z1}{A vector containing the counterfactual mean estimates under medication \eqn{Z=1}. The \eqn{k}th element in the vector corresponds to the \eqn{k}th follow-up time interval.}
}
\description{
This function applies the pooled inverse probability weighted approach described by McGrath et al. (in preparation).
}
\details{
Additional description of the method
}
